+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   * Added verification to ensure that `akamai_property_rules_builder` data source
     has consistent frozen rule format between parent and it's child.
     Additionally `akamai_property_rules_builder.json` is returning artificial field `_ruleFormat_`.
+  * Suppressed trigger of new activation for `note` field change in `akamai_property_activation` and `akamai_property_include_activation` resources.
 
 ## 5.0.1 (Jul 12, 2023)